requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7
openai==1.6.1
requests==2.31.0
python-dotenv==1.0.0
numpy==1.26.2
Werkzeug==3.0.1
